requirements.txt
# Nexus Trading System - Requirements
# Production-ready Python trading system dependencies

# Core Dependencies
numpy>=1.21.0
pandas>=1.5.0
pyyaml>=6.0
python-dotenv>=1.0.0
asyncio>=3.9.0

# Database
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
alembic>=1.8.0

# Web Framework
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-decouple>=3.8

# Technical Analysis & Machine Learning
scikit-learn>=1.2.0
torch>=1.12.0
talib-binary>=0.0.19

# Data Processing
openpyxl>=3.0.10
tqdm>=4.64.0

# Monitoring & Performance
psutil>=5.8.0
prometheus-client>=0.14.1

# Development & Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Email & Communication
jinja2>=3.1.0
boto3>=1.26.0
stripe>=5.0.0

# HTTP Clients
httpx>=0.23.0
requests>=2.28.0

# Date & Time
python-dateutil>=2.8.0

# Math & Statistics
scipy>=1.9.0

# Logging
colorlog>=6.6.0

# Utilities
click>=8.0.0
rich>=12.6.0

# MetaTrader 5 (Windows only - comment out on Linux/Mac)
# MetaTrader5>=5.0.40
